pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.0.5"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
]

[tool.coverage.run]
omit = [
    "*/setup.py",
    "*/tests/*",
    "*/build/*",
    "*/dist/*",
    "*/.tox/*",
    "*/.egg/*",
]

[tool.isort]
force_sort_within_sections = false
known_first_party = "prefmonitor"
profile = "black"

[tool.mypy]
disallow_subclassing_any = "False"
exclude = "^.*tests.*$"
ignore_missing_imports = "True"
implicit_reexport = "True"
strict = "True"

[tool.poetry]
authors = [
    "Jason Kratzer <jkratzer@mozilla.com>"
]
description = "Prefpicker bug dependency monitor"
homepage = "https://github.com/MozillaSecurity/prefpicker-monitor"
license = "MPL-2.0"
maintainers = [
    "Mozilla Fuzzing Team <fuzzing@mozilla.com>"
]
name = "prefmonitor"
readme = "README.md"
repository = "https://github.com/MozillaSecurity/prefpicker"
version = "0.1.0"

[tool.poetry.dependencies]
PyGithub = "^1.55"
bugsy = "^0.12.0"
prefpicker = {git = "https://github.com/MozillaSecurity/prefpicker.git"}
python = "^3.7"
six = "^1.16.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
codecov = "^2.1.12"
flake8 = "^4.0.1"
flake8-isort = "^4.1.1"
flake8-quotes = "^3.3.1"
gitlint = "^0.17.0"
mypy = "^0.942"
pre-commit = "^2.18.1"
pylint = "^2.13.5"
pytest = "^7.1.1"
pytest-black = "^0.3.12"
pytest-cov = "^3.0.0"
pytest-flake8 = "^1.1.1"
pytest-mock = "^3.7.0"
pytest-pylint = "^0.18.0"
python-semantic-release = "^7.28.1"
taskcluster-yml-validator = "^0.0.9"
toml = "^0.10.2"
toml-sort = "^0.20.0"
tox = "^3.25.0"

[tool.poetry.scripts]
prefmonitor = "prefmonitor.main:main"

[tool.pylint.MASTER]
load-plugins = "pylint.extensions.docparams,pylint.extensions.docstyle"

[tool.pylint.format]
max-line-length = 88

[tool.pylint.messages_control]
disable = [
    "bad-continuation",
    "fixme",
    "invalid-name",
    "line-too-long",
    "logging-format-interpolation",
    "logging-fstring-interpolation",
    "missing-module-docstring",
    "missing-return-doc",
    "missing-yield-doc",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-lines",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-return-statements",
    "too-many-statements",
    "useless-object-inheritance"
]

[tool.pytest.ini_options]
log_level = "DEBUG"

[tool.semantic_release]
branch = "main"
build_command = "pip install poetry && poetry build"
commit_subject = "chore(release): {version} [skip ci]"
upload_to_pypi = false
upload_to_release = false
version_toml = "pyproject.toml:tool.poetry.version"
